made the ui better
