--- cpp.cpp
@@ -1,11 +1,0 @@
-#include<iostream>
-#include<string.h>
-using namespace std;
-int main()
-{
-    string a="df";
-    if(a[1]=='f'){
-            cout << "das" << endl;
-
-    }
-}
