--- pr.html
@@ -1,48 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Checkbox Animation</title>
-    <style>
-        /* Hide the default checkbox */
-        #check {
-            appearance: none;
-            width: 50px;
-            height: 50px;
-            border: 2px solid black;
-            border-radius: 5px;
-            cursor: pointer;
-            position: relative;
-            transition: background-color 0.3s ease-in-out;
-        }
-
-        /* Change background when checked */
-        #check:checked {
-            background-color: green;
-        }
-        
-        /* Add tick mark when checked */
-        #check:checked::after {
-            content: "✔";
-            color: white;
-            font-size: 30px;
-            font-weight: bold;
-            position: absolute;
-            top: 50%;
-            left: 50%;
-            transform: translate(-50%, -50%) scale(0);
-            animation: pop 0.3s ease-in-out forwards;
-        }
-
-        /* Keyframe animation for the tick */
-        @keyframes pop {
-            0% { transform: translate(-50%, -50%) scale(0); }
-            100% { transform: translate(-50%, -50%) scale(1); }
-        }
-    </style>
-</head>
-<body>
-    <input id="check" type="checkbox">
-</body>
-</html>
